Converting unsigned long to size_t.

+++ torchvision/csrc/io/image/cpu/encode_jpeg.cpp
@@ -22,7 +22,7 @@ torch::Tensor encode_jpeg(const torch::Tensor& data, int64_t quality) {
   struct torch_jpeg_error_mgr jerr;
 
   // Define buffer to write JPEG information to and its size
-  unsigned long jpegSize = 0;
+  size_t jpegSize = 0;
   uint8_t* jpegBuf = NULL;
 
   cinfo.err = jpeg_std_error(&jerr.pub);
